Create utils module for utility functions
