+++ src/Components/Resources.jsx
@@ -1,17 +1,5 @@
 import { Col, Dropdown, Row } from "react-bootstrap";
-
-// src https://stackoverflow.com/questions/32589197/how-can-i-capitalize-the-first-letter-of-each-word-in-a-string-using-javascript
-function titleCase(str) {
-  var splitStr = str.toLowerCase().split(" ");
-  for (var i = 0; i < splitStr.length; i++) {
-    // You do not need to check if i is larger than splitStr length, as your for does that for you
-    // Assign it back to the array
-    splitStr[i] =
-      splitStr[i].charAt(0).toUpperCase() + splitStr[i].substring(1);
-  }
-  // Directly return the joined string
-  return splitStr.join(" ");
-}
+import { titleCase } from "../utils";
 
 const redCardsPrintLinks = {
   arabic:
